requirements.txt
google-genai
yfinance
python-dotenv
elevenlabs
flask-socketio 
eventlet